% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{adaptModel}
\alias{adaptModel}
\title{Fit the appropriate model}
\usage{
adaptModel(
  protDat,
  covType,
  fixedForm,
  bridgeDat,
  timeParm,
  lhtList,
  fullColumns,
  minRE,
  randID,
  reducedMod,
  multiLevel
)
}
\arguments{
\item{protDat}{A dataframe with all the data needed to model
a single protein.}

\item{covType}{A string defining the type of
 reference covariate.  Acceptable types are "Factor",
"Continuous", "Time" or "None.  "None" specifies the
basic model for estimating each sample}

\item{bridgeDat}{The subset of data that comes from bridge
channels}

\item{fullColumns}{The name of every variable that we
would estimate if there were no missing values}

\item{minRE}{The number of levels of a random variable required to fit a 
random intercept}

\item{tCatIndex}{An integer denoting the position in
covString of a categorical variable that should have
unique time trajectories for each level.}
}
\description{
Fit the appropriate model
}
