% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileConverter.R
\name{convertPdFile}
\alias{convertPdFile}
\title{Convert proteome discoverer ouput to R}
\usage{
convertPdFile(pd_psm_file, pd_protein_file = "", delimiter = "\\t")
}
\arguments{
\item{pd_psm_file}{A tab-delimited txt file from protoeme discoverer's 
PSMs output. It should contains Abundance columns as well as "Average Reporter S/N".
A example file pd_example_export_PSMs.txt is available in data folder.}

\item{pd_protein_file}{A tab-delimited txt file from protoeme discoverer's 
Proteins output. The file is optional, only used for getting gene name for each protein. 
It must contains "Gene.Symbol" and "Accession" columns}

\item{delimiter}{The delimiter for the files. For tsv, use "\t"; for csv, use ",".}
}
\value{
A r dataframe of the input file, or a string start with "Error:" containing 
the error message. The csv file correspondings to the df will also be save in the folder.
}
\description{
Convert proteome discoverer ouput to R
}
\examples{
df <- convertPdFile("data/pd_example_export_PSMs.txt", "data/pd_example_export_Proteins.txt")
}
