% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{findOutliers}
\alias{findOutliers}
\title{Remove "consensus outliers" from protein groupings}
\usage{
findOutliers(protMat, snMat, outlierCutoff = 3, scaleSN)
}
\arguments{
\item{protMat}{A matrix of reporter ion intensities from a
single protein.}

\item{outlierCutoff}{Outliers defined by consensus ratios
within each protein grouping will be flagged whenever
the jacknife residuals exceed this value.  Flagged scans
will be removed from the protein grouping, but will
remain in the analysis by replacing the protein label
with a peptide sequence.  Set this parameter to NULL to
skip this form of outlier removal.}

\item{scaleSN}{The technical variance of an observation is modeled
as the inverse of the square root of each SNR * some constant.
scaleSN is that constant.  This value will be used for both
imputation and modeling.}
}
\value{
A matrix of boolean values indicating which entries
 met the definition of an outlier.
}
\description{
Remove "consensus outliers" from protein groupings
}
