% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{geoNorm}
\alias{geoNorm}
\title{Perform global column adjustments}
\usage{
geoNorm(mat, normIndex, Plex, ratios)
}
\arguments{
\item{mat}{A matrix of reporter ion intensities that need
to be adjusted}

\item{normIndex}{A boolean vector with length equal to the
number of rows in "mat". Each "1" indicates that the
corresponding row will be used to calculate the normalization
factors.}

\item{Plex}{A vector denoting plex membership.  Normalization
factors are computed for every channel*plex combination.
Accordingly, this variable determines how many adjustments
will be made.}

\item{ratios}{A numeric vector with length equal to the
number of normalization factors.  The ratios in this
vector determine the expected relationship between
all of the columns.  Typically it will be a vector of 1's,
but occasionally experiments will include columns at known
dilutions, e.g. a 2x bridge channel.}
}
\value{
A length two list.  the first component includes a
 matrix that has been adjusted so that the geometric mean
 of the subset of rows specified by "normIndex" will be
 equivalent across all columns.  The second list component
 contains a vector with all the log additive values used
 to adjust each column.
}
\description{
Perform global column adjustments
}
