% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{miniTrawl}
\alias{miniTrawl}
\title{Function for modeling rdata created by the protPrep function.
The function is intended to be used as part of a parallel
processing procedure.}
\usage{
miniTrawl(rdaSubset)
}
\arguments{
\item{rdaSubset}{An rda datafile created by protPrep.  This is a
subset of the data and all the relevant parameters needed to
fit models and generate results.}
}
\description{
Function for modeling rdata created by the protPrep function.
The function is intended to be used as part of a parallel
processing procedure.
}
