% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{protPrep}
\alias{protPrep}
\title{Prepare protein subsets for parallel analysis}
\usage{
protPrep(
  DF,
  setSize = 50,
  sampleFile = NULL,
  covariateFile = NULL,
  scaleSN = 1,
  lod = 0.01,
  imputePenalty = 1,
  minAbove = 3,
  ssnFilter = 20,
  outlierCutoff = 3,
  N_SUM = 3,
  swapProtein = FALSE,
  maxPep = 25,
  colAdjust = 0.5,
  colRatios = NULL,
  dropContam = TRUE,
  dropReverse = TRUE,
  peptideAnalysis = FALSE,
  minRE = 5,
  timeDiff = TRUE
)
}
\arguments{
\item{DF}{A dataframe containing identifiers and
intensities that matches the structure in the sample
data.}

\item{setSize}{An integer value denoting the approximate
number of proteins that will be included in each
subset of data.
if setSize == 50 and the number of proteins is 99,
    there will be one set of size 99.
If setSize == 50 and the number of proteins is 100,
    there will be two sets of size 50.
If set Size == 50 and the number of proteins is 101,
    there will be two set, one of size 50, the other of size 51.}

\item{sampleFile}{A csv file containing all the covariate
values that map to individual samples in the study. An
example file called sampleFile can be found in the data
folder.}

\item{covariateFile}{A csv file containing all covariate
metadata that is used to determine the type of analyses
that will be performed.  See covariateFile in the data
folder.}

\item{scaleSN}{The technical variance of an observation is modeled
as the inverse of the square root of each SNR * some constant.
scaleSN is that constant.  This value will be used for both
imputation and modeling.}

\item{lod}{A value that determines the limit of reliability.
If the value is between zero and one, then
the threshold will be a proportion of the total scan
intensity.  If lod >= 1, then the threshold will
be set at a SNR = lod for all scans.}

\item{minAbove}{It is possible that all of the scans will
be below the limit of reliability.  Imputing an entire scan
is senseless.  For small numbers of values above lod the
decision is questionable.  If fewer than minAbove observations
are above the lod, then the scan will be flagged for
removal.}

\item{ssnFilter}{A cutoff for each scan based on the sum of
signal-to-noise ratios across all channels.  This filter
precedes all others, so if the a scan has less total signal
than "ssnFilter", then no LOD imputation will occur.  If
NULL then no SSN filtering will be done.}

\item{outlierCutoff}{Outliers defined by consensus ratios
within each protein grouping will be flagged whenever
the jacknife residuals exceed this value.  Flagged scans
will be removed from the protein grouping, but will
remain in the analysis by replacing the protein label
with a peptide sequence.  Set this parameter to NULL to
skip this form of outlier removal.}

\item{N_SUM}{If the number of observations from a protein
within a sample, is less than N_SUM, then both the fluxes
and corresponding ion counts will be aggregated into a single
data point.  Setting this to a high value will activate the
single-level msTrawler model for all proteins.}

\item{swapProtein}{Boolean variable that determines the
handling of outlier scans.  When a scan includes an observation
that exceeds the outlierCutoff, we will either remove the scan
from the analysis, or replace the protein label with the
corresponding peptide label.  This lets us look for relationships
that may be indicative of interesting post translational modifications.
When swapProtein == TRUE, we perform the label swap.  Otherwise,
we remove the scan.}

\item{maxPep}{An integer denoting the maximum number of scans
to use from a single sample.  In plasma many proteins have
hundreds of scans that can result in an out of memory error
when working with large experiments.  Only the scans with the
highest summed-signal-to-noise ratios will be used in the modeling.}

\item{colAdjust}{This parameter specifies how global column
adjustments will be performed.  The parameter can
take 3 types of input.  A value between 0 and 1
will be interpreted as a standard deviation percentile.
Only rows with a standard deviation percentile less than
the entered value will be used to calculate column
adjustment factors.  If colAdjust = NULL, then no
column adjustment will be done at all.  The last
type of acceptable parameter entry is a boolean
vector with length equal to the number of rows in DF.
Only the rows corresponding to 1s in this vector will
be used to calculate column adjustment factors.}

\item{colRatios}{A numeric vector with length equal to the
number of normalization factors.  The ratios in this
vector determine the expected relationship between
all of the columns.  Typically it will be a vector of 1's,
but occasionally experiments will include columns at known
dilutions, e.g. a 2x bridge channel.}

\item{dropContam}{Boolean variable that determines whether or
not scans containing the string "contaminant" within
the protein name will be removed prior to analysis.}

\item{dropReverse}{Boolean variable that determines whether or
not scans containing the string "##" within
the protein name will be removed prior to analysis.}

\item{peptideAnalysis}{A boolean parameter set to false by
default.  When TRUE, protein labels will be replaced with
peptide the peptide labels prior to calling any other functions.
In words, we will estimate relationships for peptides instead of
proteins.  This also triggers the single-sample msTrawler model
regardless of the number of scans observed per peptide.}

\item{minRE}{The number of levels required to fit a random intercept.}

\item{timeDiff}{A boolean parameter that determines whether or
not hypothesis tests for differential time trends will occur.
The only reason to set this to false is if computational time
is a concern and trends across categories are not of interest.}
}
\description{
Prepare protein subsets for parallel analysis
}
