% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{tmtLOD}
\alias{tmtLOD}
\title{Impute LOD}
\usage{
tmtLOD(snMat, iMat, lod = 0.01, minAbove = 3, scaleSN, imputePenalty)
}
\arguments{
\item{snMat}{A matrix of signal-to-noise ratios with
channels in the columns and scans in each row}

\item{iMat}{A matrix of intensities with
channels in the columns and scans in each row}

\item{lod}{A value that determines the limit of reliability.
If the value is between zero and one, then
the threshold will be a proportion of the total scan
intensity.  If lod >= 1, then the threshold will
be set at a SNR = lod for all scans.}

\item{minAbove}{It is possible that all of the scans will
be below the limit of reliability.  Imputing an entire scan
is senseless.  For small numbers of values above lod the
decision is questionable.  If fewer than minAbove observations
are above the lod, then the scan will be flagged for
removal.}

\item{scaleSN}{The technical variance of an observation is modeled
as the inverse of the square root of each SNR * some constant.
scaleSN is that constant.  This value will be used for both
imputation and modeling.}
}
\value{
A length three list. This first component has
 a matrix of post imputation intensities and the second
 has the post imputation SNRs.  The final component contains
 an indicator variable for flagging scans with too few
 observations above the lod (as determined by minAbove).
}
\description{
Impute LOD
}
